requirements.txt
python-dateutil
pyyaml
websockets
aiohttp
